refactor: 抽离时空域部分函数为外域公共函数类 foreignDomainUtils
