+++ agent/agent_allfile.py
@@ -1,6 +1,7 @@
 from action.fight import jjc101
 from action.fight import getKeyFromHole
 from action.fight import timeSpaceDomain
+from action.fight import foreignDomainUtils
 from action.activity import kairo_maze1
 from action.activity import eastern
